Use correct blinding and base for D
The value D in the Ped-DLEQ proof is a blinding
of the user's chosen pubkey. To tie the PedDLEQ and the curve tree set
membership proofs together, we need to use the same blinding in this
case.
In this commit we extract the random scalar and the blinding basepoint
from the select-and-rerandomize algorithm execution, to be inserted into
the PedDLEQ proof.
To be checked: is it correct to use the even parameters?