print out tree root in prover and verifier
.. confirming that the proof is for the same tree, it's debatable if
this is needed: if the Verifier constructs the tree from the keyset, and
the provided proof verifies against it, then it's implicit that it must
be correct, but it should be a very useful sanity check to be able to
compare the Prover's proof data with the reconstructed root.
Unfortunately the codebase makes it a little difficult to do this
elegantly (? I think).
Also updated the key image value which was previously wrong because we
didn't use little endian encoding for the private key.