Basic RPC architecture: verify as call
The server prepares the CurveTree and SelRerandParams objects, allowing
a remote client to call verify() and have the result returned quickly
(testing gives between 50 and 120ms for this base case).
Negative integers are returned for verification failures, 1 is returned
for a fully valid proof.
New binaries rpcserver and rpcclient created for this purpose. The autct
binary is still present for creating proofs.