Corrected dealing with non-permissible pubkeys
Prior to this commit, some (most) keys would not produce valid proofs
in that the rerandomized key from the Curve Tree proof did not match the
Ped-DLEQ proof. The reason for this is that, if the starting pubkey was
not permissible and was therefore automatically incremented by a small
multiple of the basepoint B_blinding, the randomness returned and used
in the Pedersen commitment to P (called D in this codebase), was not the
actual full randomness. We now increment the randomness to match, so
proofs will verify for any starting pubkey.
Additionally, we address the possibility that the pubkeys generated by
the BIP340 algorithm, for a specific private key, use a tiebreaker
algorithm different from that in ark-ec; we do this by flipping the sign
of the private key if necessary.