Command line options with help for binaries
This is a big overhaul to make running the commands from the terminal
intuitive and in line with what a user would expect. The priority is:
command line options overwite contents of config file, which overwrite
defaults. The config is still persisted with confy, and clap is used to
print out useful command line help and manage option parsing.

The examples in the README are updated to reflect this.
One minor point un-addressed: some binaries require more than just the
`keyset` option, but the help message does not mention this.