Added keyimagestore module to track key image uses
This is a first draft, and the logic in the verifier to reject reuse has
not yet been implemented.
Also removed the keyfinder module as it's not need for the main (utxo)
use case and has a lot of problems to resolve before it's usable (it
would be used to find keysets of spent keys from block ranges).