Context label in keyimagestore filename
Prior to this commit, the keyimage store filename was defined only by
the app domain label, which is clearly incorrect; we need different key
image stores for each context, so the filename now includes the context
label also. Also fixed up the tests to create separate testing files and
delete them on completion.