Support multiple contexts for server
After this commit, contexts are set with keysets in the --keyset option
or config var, as pairs in a comma separated list, split by a ":". This
is used by the "serve" function to allow simultaneous verification of
keysets for multiple contexts. To support this, we now create one
KeyImageStore object for each of the contexts that are active. Note that
multiple contexts are of course not allowed for the prove or request
function.