Adds method to generate taproot key and address
Due to limitations of existing wallets, the easiest way currently for a
user to generate a taproot utxo is to generate the private key in this
tool (using the standard way recommended in rust-bitcoin), print the WIF
encoding of this private key, along with the address for the configured
network.
This gives them a one step process to get both those strings, use the
address to fund the utxo, and use the WIF as input to the autct prover,
as well as using the same WIF as input to the sweep/import function in
external wallets to access the funds at a later time.

The new mode is -M newkeys and just prints the new WIF and p2tr address,
and exits.