Convert to permissible in method "convertkeys"
Fixes #10.
This patch makes the actual proving time faster by 80% or more, and
addresses the problem described in #10, namely that the time for a
client to create a proof was dominated by the conversion of all the
pubkeys in the set to permissible curve points.
To achieve this goal, a new method "convertkeys" is added, which takes
in a set of hex encoded raw pubkeys, and outputs a binary serialization
of the permissible equivalent of each point, in a file named *.aks.p .

Hence the "prove" method now takes this new binary serialization as
input and just assumes that all the points are already permissible, thus
being able to construct the CurveTree object directly.