Make point deserialization unchecked
Removing checks reduces the time for deserialization of 330K keys from
over 100s to about 10s. Obviously this is the main slowdown culprit but
there may be more to be done to improve this further.
Meanwhile this is a waypoint commit ... it still has some timing code
and the readme/examples still need updating for the new keyset format.