Migrate all functions to rpc calls
To simplify the user interface and make further development easier, all
functions provided by the application will be possible as RPC calls. In
this commit, only the proving and the verifying are done in this way.
Additionally, to streamline the process, the convertkeys function is
removed as it is only needed for proving and verifying and therefore can
be done on startup of the RPC server, removing any delays during the
proving step.
What remains to be done is: add newkeys to the RPC interface, and,
importantly, introduce security in the form of encryption of the file
holding the private key which is required for proving.