Bugfix the randomness offset for permissibility
Prior to this commit some verifications failed, due to incorrect
algorithm for calculating the randomness offset in the ped-dleq proof to
correctly match the curve tree proof.
This happened because in the refactoring to allow pre-computation of the
permissible point versions of each of the real taproot pubkeys (and
stored in the .aks.p file, see the convert_keys method), I forgot to
change the algorithm to re-derive that number of H bumps that were used.
As a reminder, the issue at hand is:
In order for the Ped-DLEQ proof to "match" the Curve Tree proof, the
blinding factor r applied in the Pedersen commitment (xG +rH) in the
Ped-DLEQ proof must be the same as the blinding factor r in the curve
tree proof (where you produce a blinded version of the pubkey and prove
it is part of the tree). To make those values the same, the Ped-DLEQ
part has to figure out what r to use, and it is not just "use the same
blinding factor" but "use the same blinding factor, with a number of +H
added to match what is needed for the curve point to be permissible").