RPC client-server over websockets
This change does not affect CLI usage of the client and server, but is
done to make accessing the API via clients of other
languages/applications easer.
See https://github.com/AdamISZ/autct-api for an example client in
Python.

An additional small change is that the verify RPC call now submits the
proof in base64 encoding, not binary, to rationalize types used in the
RPC exchange.