Handle errors from external caller data
This is a first round of "cleanup" of the codebase to make it so that
uses of unwrap(), expect(), assert! and similar are restricted only to
situations in which the error can only be triggered by an internal logic
failure.
Thus, errors from user input (or external input in the case of the RPC
server) no longer cause crashes. The focus has been on the two RPC calls
prove and verify; some more work may well be needed, on the createkeys
function and a few other places.
As part of this, several new error codes were defined for the prove and
verify calls which can affect downstream usage.