Add encryption and decryption of private key files.
Uses AES-GCM with Argon for hashing/salting passwords.
User must provide password for encrypted private key file both when
creating a new private key, and also when requesting a proof. A method
encryptkey is added to the CLI to allow encrypting an existing WIF read
from a file, and writing the now encrypted WIF key to a new file with
file suffix ".enc" appended.
The README worked example is updated to include this encryption.

cleanup error handling and modify README