Implement zk proofs of total coin ownership in range
This serves the function of auditing with privacy:
given a (large) set of candidate utxos, prove, using
a combination of curve tree proofs of set membership,
bulletproofs of sum in range and generalised schnorr
proofs of representation, that the sum of the utxos'
value is in a given range without revealing those utxos,
while preventing duplicate use of utxos.

The range of the sum is defined as between k and
k + 2^n for specified integers k, n.
The utxos are represented as commitments P + vJ where
v is the value of the utxo in sats, P is the taproot
pubkey of the utxo, and k < sum(v) < k + 2^n.