GUI now starts verif server in background
This is done by user clicking a dot which changes color.
The server's slow startup is monitored with a new RPC
call 'echo'.
The app is now basically fully functional, but rough edges
need cleaning off, in particular the syncing of the config
vars that are being set in the window, with those used
to start up the server.