Updated for auditing function, performance improvements
The auditor now reads .pks format files, which are ark-ec
but in hex, whitespace delimited.
Performance of proving is very greatly improved, 1-2 second
proving time for a tree of 2-500K size. This is from removing
redudant deserialization which was only being used to do an index
find; this can easily be done on the hex strings instead of
deserialized points.
This version should now be functionally equivalent to that in
the master branch.